+++ src/Component/TopNabar.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function TopNavbar() {
+  return (
+    <div className='bg-black text-white text-center font-serif p-5'>
+        FREE SHIPPING ALL INDIA
+    </div>
+  )
+}
